projects/sql_dlt/pyproject.toml
[project]
name = "sql_dlt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-dlt>=0.3.0",
    "pyspark>=3.5.5",
]

[dependency-groups]
dev = [
    "pip>=25.0.1",
    "pytest>=8.3.5",
    "ruff>=0.9.10",
]
